WIP: Research for flickering on resize
When resizing CropRect with one of the CropRectResizeHandler, a
flickering appears.
Assumably, this is being caused by the `updateCropRect` function being called 2
times per event - one with the correct values, the second time with
different ones. This could be an issue with react-draggable internal
state handling.
Using DraggableCore from react-draggable, which should disable internal
state handling results in the same error.